requirements.txt
streamlit
langchain
openai
tiktoken
pypdf
pinecone-client
PyPDF2
pdfplumber
openpyxl